refactor: 비교 결과 페이지 컴포넌트 구조 개선 및 FSD 레이어 재정리
- 비교 결과 탭을 `CompareCompleteTabs` 위젯으로 추출
- 탭 관련 UI 컴포넌트 `widgets/compare-complete-tabs`로 이동
- 공통 타입(`CellData`, `RowData`, `SlideProps`, `TableProps`)을 `entities/compare`로 이동
