+++ apps/knockdog/src/entities/compare/model/types.ts
@@ -27,6 +27,10 @@ export type {
   DistanceComparisonsByTransport,
   DistanceDetailComparison,
   ShortestInfo,
+  CellData,
+  RowData,
+  SlideProps,
+  TableProps,
 } from './compare-result';
 
 interface KindergartenShortInfo {
