+++ apps/knockdog/src/features/compare/index.ts
@@ -18,19 +18,12 @@ export { createOperatingScheduleSlide } from './lib/createOperatingScheduleSlide
 
 /** ui */
 export { PricingSummary } from './ui/PricingSummary';
-export { PricingSection } from './ui/PricingSection';
 export { PriceDetailedItem } from './ui/PriceDetailedItem';
 
 export { DistanceSummary } from './ui/DistanceSummary';
-export { DistanceSection } from './ui/DistanceSection';
 export { DistanceDetailedItem } from './ui/DistanceDetailedItem';
-
-export { PickdropSection } from './ui/PickdropSection';
-export { HolidaySection } from './ui/HolidaySection';
-export { OperatingDaysSection } from './ui/OperatingDaysSection';
 
 export { ComparisonSimpleItem } from './ui/ComparisonSimpleItem';
 export { ComparisonDaysItem } from './ui/ComparisonDaysItem';
 
-export { DogServiceSection } from './ui/DogServiceSection';
 export { ComparisonHistoryCard } from './ui/ComparisonHistoryCard';
