+++ apps/knockdog/src/views/compare-complete-page/ui/CompareCompletePage.tsx
@@ -1,35 +1,14 @@
 'use client';
 
-import { PropsWithChildren, useMemo } from 'react';
+import { useMemo } from 'react';
 import { useSearchParams } from 'next/navigation';
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '@knockdog/ui';
 import { Header } from '@widgets/Header';
-import { SwipeCarousel } from '@widgets/comparisons-tab/ui/SwipeCarousel';
-import { Table } from '@widgets/comparisons-tab/ui/Table';
-import {
-  useComparisonsQuery,
-  PricingSection,
-  DistanceSection,
-  PickdropSection,
-  HolidaySection,
-  OperatingDaysSection,
-  createPricingSlides,
-  createDistanceSlides,
-  createOperatingScheduleSlide,
-  DogServiceSection,
-} from '@features/compare';
+import { CompareCompleteTabs } from '@widgets/compare-complete-tabs';
+import { useComparisonsQuery } from '@features/compare';
 import type { KindergartenComparison } from '@entities/compare';
 import { SelectedCell, serializeCategories, resolveIds, s3ToUrl } from '@entities/compare';
 import { SafeArea } from '@shared/ui/safe-area';
 import { useShare } from '@shared/lib/device/useShare';
-
-function ComparisonPanel({ children }: PropsWithChildren) {
-  return <div className='bg-fill-secondary-0 rounded-2xl py-10'>{children}</div>;
-}
-
-function ComparisonSection({ children }: PropsWithChildren) {
-  return <section className='border-line-200 border-b px-4 last:border-b-0'>{children}</section>;
-}
 
 function CompareCompletePage() {
   const params = useSearchParams();
@@ -44,10 +23,6 @@ function CompareCompletePage() {
   const [left, right] = useMemo(() => data?.filter((item): item is KindergartenComparison => !!item) ?? [], [data]);
 
   if (!left || !right) return null;
-
-  const pricingSlidesData = createPricingSlides(left, right);
-  const distanceSlidesData = createDistanceSlides(left, right);
-  const operatingSlideData = createOperatingScheduleSlide(left, right);
 
   const handleShare = () => {
     const url = `https://knockdog.com/compare-complete?ids=${left.id},${right.id}`;
@@ -94,49 +69,7 @@ function CompareCompletePage() {
           })}
         </div>
 
-        <Tabs defaultValue='summary' className='flex flex-col overflow-hidden'>
-          <TabsList>
-            <TabsTrigger value='summary'>요약</TabsTrigger>
-            <TabsTrigger value='details'>자세히</TabsTrigger>
-          </TabsList>
-          <TabsContent value='summary' className='overflow-y-auto'>
-            <div className='bg-text-primary min-h-full space-y-3 px-4 py-7'>
-              <ComparisonPanel>
-                <ComparisonSection>
-                  <PricingSection left={left} right={right} />
-                </ComparisonSection>
-                <ComparisonSection>
-                  <PickdropSection left={left} right={right} />
-                </ComparisonSection>
-              </ComparisonPanel>
-
-              <ComparisonPanel>
-                <ComparisonSection>
-                  <DistanceSection left={left} right={right} />
-                </ComparisonSection>
-              </ComparisonPanel>
-              <ComparisonPanel>
-                <ComparisonSection>
-                  <OperatingDaysSection left={left} right={right} />
-                </ComparisonSection>
-                <ComparisonSection>
-                  <HolidaySection left={left} right={right} />
-                </ComparisonSection>
-              </ComparisonPanel>
-            </div>
-          </TabsContent>
-          <TabsContent value='details' className='overflow-y-auto'>
-            <div className='flex flex-col gap-5 px-4 py-7'>
-              <SwipeCarousel title='요금 비교' slides={pricingSlidesData} />
-
-              <SwipeCarousel title='거리 비교' slides={distanceSlidesData} />
-
-              <DogServiceSection leftService={left.service} rightService={right.service} />
-
-              <Table title='운영 시간 비교' rows={operatingSlideData} />
-            </div>
-          </TabsContent>
-        </Tabs>
+        <CompareCompleteTabs left={left} right={right} />
       </div>
     </SafeArea>
   );
