+++ apps/knockdog/src/widgets/compare-complete-tabs/index.ts
@@ -1,0 +1,1 @@
+export { CompareCompleteTabs } from './ui/CompareCompleteTabs';
