+++ apps/knockdog/src/widgets/compare-complete-tabs/ui/ComparisonSection.tsx
@@ -1,0 +1,7 @@
+import { PropsWithChildren } from 'react';
+
+function ComparisonSection({ children }: PropsWithChildren) {
+  return <section className='border-line-200 border-b px-4 last:border-b-0'>{children}</section>;
+}
+
+export { ComparisonSection };
