+++ apps/knockdog/src/widgets/compare-complete-tabs/ui/SwipeCarousel.tsx
