+++ apps/knockdog/src/widgets/compare-complete-tabs/ui/Title.tsx
